+++ src/main/java/beachcombing/backend/domain/feed/Feed.java
@@ -1,0 +1,18 @@
+package beachcombing.backend.domain.feed;
+import jakarta.persistence.*;
+import lombok.Getter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+public class Feed {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "feed_id")
+    private Long id;
+
+    private String review;
+
+}
